Continuando a implementação do serviço /automoveis

+++ src/main/java/core/entity/TipoDespesaDiversa.kt
@@ -13,5 +13,5 @@ open class TipoDespesaDiversa {
     var id: UUID? = null
 
     @Version
-    val versao: Long? = null
+    val versao: Int? = null
 }
